ml/requirements.txt
fastapi~=0.111.0
pydantic~=2.7.1
uvicorn~=0.29.0
cohere~=5.3.4
python-dotenv~=1.0.1
minio~=7.2.7
requests~=2.31.0
transformers~=4.41.0
torch~=2.0.1
datasets~=2.19.1